requirements.txt
asgiref==3.2.10
cffi==1.14.3
cryptography==3.1.1
Django~=3.1.4
django-rename-app==0.1.2
gunicorn==20.0.4
jwcrypto==0.8
Pillow==7.2.0
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.2.0

djangorestframework~=3.12.2
pip~=20.3.3
django-filter~=2.4.0